% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSRRFastqsToGSMs.R
\name{combineSRRFastqsToGSMs}
\alias{combineSRRFastqsToGSMs}
\title{This function takes a data frame with information on SRR IDs and the state of
the download of their associated .fastq files from the European Nucleotide
Archive, such as one generated by the 'SRRFastqDownload' function, and
combines the .fastq files associated with the SRRs in a user-selected
directory into larger .fastq files associated with each GSM ID that called
the individual SRRs.}
\usage{
combineSRRFastqsToGSMs(
  SRRFastqDownloadDF,
  localDownloadDirectory = getwd(),
  outputGSMFastqDirectory = getwd()
)
}
\arguments{
\item{SRRFastqDownloadDF}{Specify a dataframe with relevant information
about the SRRs and their associated .fastq files, such as one output by the
SRRFastqDownload function. This dataframe should contain columns with
the names 'GSM_IDs', 'SRR_IDs', 'Sequencing_type', and 'Fastq_file_sizes'.}

\item{localDownloadDirectory}{Set a path to a directory where the user has
downloaded the .fastq files associated with the SRRs detailed in the
\code{SRRFastqDownloadDF}. Defaults to the user's working directory.}

\item{outputGSMFastqDirectory}{Set a path to a directory where the user wants
the combined .fastq files for the GSM IDs to be output to. Defaults to the
user's working directory.}
}
\value{
Combines the previously downloaded SRR .fastq files in the directory
specified as the \code{localDownloadDirectory} into .fastq files for each GSM ID.
These GSM IDs will be deposited in the directory  The function will also return the dataframe
given as the \code{identifySRRsFromGSMIDsDF} argument with an additional column
named "SRR_file_download_status" noting the status of the download of each
.fastq file.
}
\description{
This function takes a data frame with information on SRR IDs and the state of
the download of their associated .fastq files from the European Nucleotide
Archive, such as one generated by the 'SRRFastqDownload' function, and
combines the .fastq files associated with the SRRs in a user-selected
directory into larger .fastq files associated with each GSM ID that called
the individual SRRs.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## This example will attempt to combine the 6? .fastq files for SRRs
## associated with three example GSM IDs using an example dataset with
## results from the previous 'SRRFastqDownload' function.
## NOTE: TO RUN THIS EXAMPLE, THE EXAMPLE FROM THE 'SRRFastqDownload' MUST
## HAVE BEEN PREVIOUSLY RUN TO DOWNLOAD SRR .fastq FILES TO THE USER'S
## WORKING DIRECTORY.

## Load the example dataset.
exampleDFGEODownloadR <- utils::data(
    "exampleDFGEODownloadR",
    package = "GEODownloadR"
)

## Combine the SRR .fastqs in the user's working directory into .fastqs
##  associated with the original 3 GSM IDs.
returnValue <- combineSRRFastqsToGSMs(
    SRRFastqDownloadDF = exampleDFGEODownloadR
)
\dontshow{\}) # examplesIf}
}
