% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeGEODownloadRWorkflow.R
\name{completeGEODownloadRWorkflow}
\alias{completeGEODownloadRWorkflow}
\title{A complete workflow of GEODownloadR functions}
\usage{
completeGEODownloadRWorkflow(
  GSMIDVector,
  useCombineSRRFastqsToGSMs = TRUE,
  localDownloadDirectory = getwd(),
  outputGSMFastqDirectory = getwd(),
  downloadSRRWithGSMPresent = FALSE,
  SRRInformationNodeCount = 12,
  SRRDownloadNodeCount = 12
)
}
\arguments{
\item{GSMIDVector}{Specify a vector of sample accession numbers to
investigate/download files for. The values in the GSMIDVector should be given
in the form "GSM" followed by the numbers associated with the sample, such as
"GSM2437766".}

\item{useCombineSRRFastqsToGSMs}{Set to TRUE to use the
combineSRRFastqsToGSMs function to combine the downloaded SRR .fastq files
back into .fastq files associated with each GSM ID directly. Defaults to
TRUE.}

\item{localDownloadDirectory}{Set a path to a directory where the user wishes
to download the .fastq files associated with the SRRs. Defaults to the user's
working directory.}

\item{outputGSMFastqDirectory}{Set a path to a directory where the user wants
the combined .fastq files for the GSM IDs to be output to. Defaults to the
user's working directory.}

\item{downloadSRRWithGSMPresent}{Set to TRUE to try downloading SRR .fastq
files even when the final GSM .fastq files associated with those SRRs
(compiled by the 'combineSRRFastqsToGSMs' function) is present, otherwise set
to FALSE to not. Defaults to FALSE.}

\item{SRRInformationNodeCount}{Specify a maximum number of nodes to use when
getting SRR information (particularly the SRR file sizes). This will only
have an effect if \code{pullFastqFileSizes} is TRUE. Try setting to a relatively
low number (like 1) if the function is having difficulty in acquiring the
fastq file size (and there are many "FAILED" values in the
'Passed_SRR_file_information' column of the returned data frame). Defaults to
12.}

\item{SRRDownloadNodeCount}{Specify a maximum number of nodes to use when
downloading SRR .fastqs. Try setting to a relatively
low number (like 1) if the function is having difficulty in acquiring the
fastq files (and there are many "FAILED" values in the
'SRR_file_download_status' column of the returned data frame). Defaults to
12.}
}
\value{
First, this function finds and downloads the SRR .fastq files
associated with the user's GSM IDs given as the \code{GSMIDVector} to the
directory specified as the \code{localDownloadDirectory}, then, if
\code{useCombineSRRFastqsToGSMs} is set to TRUE, this function combines these into
.fastq files for each GSM ID as a whole. These GSM IDs will be deposited in
the directory specified as the \code{outputGSMFastqDirectory} argument. The
function will also return a data frame with information related to the GSMs,
SRR files, and the status of each function that was run.
}
\description{
This function acts as a wrapper which combines the three GEODownloadR
functions (identifySRRsFromGSMIDs, SRRFastqDownload, and optionally
combineSRRFastqsToGSMs) into a single streamlined function. This function
will also include automatic rerunning on GSM IDs that fail to download or
combine properly, and will issue a message to the user (along with the usual
dataframe of information output by either the SRRFastqDownload or
combineSRRFastqsToGSMs functions) noting the GSM IDs which have still failed
after this function's checks, so the user can more closely inspect those IDs.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## This example will attempt to pull, download, and combine information for
## three example GSM IDs and returns a data frame with that information to
## the user. Running this example requires an internet connection, and it
## will attempt to download and combine files to the user's working
## directory. It will also use 8 nodes when pulling information and
## downloading files.

returnValue <- identifySRRsFromGSMIDs(
    GSMIDVector = c("GSM4110236", "GSM2774971", "GSM2406903"),
    useCombineSRRFastqsToGSMs = TRUE,
    SRRInformationNodeCount = 8
    SRRDownloadNodeCount = 8
)
\dontshow{\}) # examplesIf}
}
