% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySRRsFromGSMIDs.R
\name{identifySRRsFromGSMIDs}
\alias{identifySRRsFromGSMIDs}
\title{Identify SRR files associated with GEO GSM IDs:}
\usage{
identifySRRsFromGSMIDs(GSMIDVector, pullFastqFileSizes = TRUE)
}
\arguments{
\item{GSMIDVector}{Specify a vector of sample accession numbers to
investigate/download files for. The values in the GSMIDVector should be given
in the form "GSM" followed by the numbers associated with the sample, such as
"GSM2437766".}

\item{pullFastqFileSizes}{Set to TRUE to pull the size of the fastq file(s)
associated with each SRR, otherwise set to FALSE to not. This option will
take awhile to run but is essential to successfully download and combine the
fastq files later. Defaults to TRUE.}
}
\value{
Returns a dataframe containing a dataframe with information about the
SRR files associated with the GSM IDs specified in the \code{GSMIDVector},
including paths to their associated .fastq files on the ENA server.
}
\description{
This function takes GSM IDs provided by the user and identifies the SRRs
associated with them. This function also provides URLs to their corresponding
.fastq files hosted by the European Nucleotide Archive (ENA). If selected,
this function will also query the ENA server to identify the size of these
.fastq files, which will take awhile to run but is essential for this
package's download pipeline. This function is the first step in the pipeline
to download and combine .fastq files associated with the user's GSM IDs of
interest.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## This example requires an internet connection, and will attempt to pull
## information, including .fastq file size, for three example GSM IDs and
## returns a data frame with that information.
returnValue <- identifySRRsFromGSMIDs(
    GSMIDVector = c("GSM4110236", "GSM2774971", "GSM2406903")
)
\dontshow{\}) # examplesIf}
}
