% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRFastqDownload.R
\name{SRRFastqDownload}
\alias{SRRFastqDownload}
\title{This function takes a data frame with information on SRR IDs and their
associated .fastq files from the European Nucleotide Archive, such as one
generated by the 'combineSRRFastqsToGSMs' function, and downloads the .fastq
files associated with them to a user-selected directory.}
\usage{
SRRFastqDownload(
  identifySRRsFromGSMIDsDF,
  localDownloadDirectory = getwd(),
  downloadSRRWithGSMPresent = FALSE
)
}
\arguments{
\item{identifySRRsFromGSMIDsDF}{Specify a dataframe with relevant information
about the SRRs and their associated .fastq files, such as one output by the
identifySRRsFromGSMIDs function. This dataframe should contain columns with
the names 'GSM_IDs', 'SRR_IDs', 'Sequencing_type', 'Fastq_URLs',
'Fastq_file_names', and 'Fastq_file_sizes'.}

\item{localDownloadDirectory}{Set a path to a directory where the user wishes
to download the .fastq files associated with the SRRs detailed in the
\code{SRRIdentificationDF}. Defaults to the user's working directory.}

\item{downloadSRRWithGSMPresent}{Set to TRUE to try downloading SRR .fastq
files even when the final GSM .fastq files associated with those SRRs
(compiled by the 'combineSRRFastqsToGSMs' function) is present, otherwise set
to FALSE to not. Defaults to FALSE.}
}
\value{
Downloads and checks the .fastq files to the directory specified as
the \code{localDownloadDirectory}. The function will also return the dataframe
given as the \code{identifySRRsFromGSMIDsDF} argument with an additional column
named "SRR_file_download_status" noting the status of the download of each
.fastq file.
}
\description{
This function takes a data frame with information on SRR IDs and their
associated .fastq files from the European Nucleotide Archive, such as one
generated by the 'combineSRRFastqsToGSMs' function, and downloads the .fastq
files associated with them to a user-selected directory.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## This example requires an internet connection, and, if run, will attempt to
## download the .fastq files for SRRs associated with three example GSM IDs
## using an example dataset with results from the previous
## 'identifySRRsFromGSMIDs' function to the user's working directory.

## Load the example dataset.
exampleDFGEODownloadR <- utils::data(
    "exampleDFGEODownloadR",
    package = "GEODownloadR"
)

## Download the SRR .fastqs. There should be a total of 11 .fastq files
## donwloaded.
returnValue <- SRRFastqDownload(
    identifySRRsFromGSMIDsDF = exampleDFGEODownloadR
)
\dontshow{\}) # examplesIf}
}
